+++ src/components/Header/index.tsx
@@ -3,7 +3,7 @@ import GitHubIcon from "@/assets/github.svg";
 import LogoIcon from "@/assets/logo.svg";
 import TwitterIcon from "@/assets/twitter.svg";
 
-function Header() {
+export function Header() {
   return (
     <header className="sticky top-0 z-50 w-full border-border/40 bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
       <div className="flex h-14 max-w-screen-2xl items-center justify-between">
@@ -56,5 +56,3 @@ function Header() {
     </header>
   );
 }
-
-export default Header;
