+++ src/components/Information/InfoDialog.tsx
@@ -11,29 +11,29 @@ import {
 } from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 
-export function WarningAlert() {
+export function InfoDialog() {
   return (
     <AlertDialog>
       <AlertDialogTrigger asChild>
         <Button
           variant="outline"
-          className="w-fit border-amber-500 text-amber-500 hover:bg-amber-100 hover:text-amber-500"
+          className="w-fit border-blue-500 text-blue-500 hover:bg-blue-100 hover:text-blue-500"
         >
-          Warning
+          Informative
         </Button>
       </AlertDialogTrigger>
       <AlertDialogContent className="ltr">
-        <div className="border-s-4 border-amber-500 pl-6">
+        <div className="border-s-4 border-blue-500 pl-6">
           <AlertDialogHeader>
-            <AlertDialogTitle>Warning Message</AlertDialogTitle>
+            <AlertDialogTitle>Info</AlertDialogTitle>
             <AlertDialogDescription>
-              This is a warning message. Please take caution before proceeding
-              further.
+              This is an informative message to notify you of something
+              important. Please review the information carefully.
             </AlertDialogDescription>
           </AlertDialogHeader>
           <AlertDialogFooter>
             <AlertDialogCancel>Dismiss</AlertDialogCancel>
-            <AlertDialogAction className="w-fit bg-amber-500 text-white hover:bg-amber-700 hover:text-white focus:ring-2 focus:ring-amber-500">
+            <AlertDialogAction className="w-fit bg-blue-500 text-white hover:bg-blue-700 hover:text-white focus:ring-2 focus:ring-blue-500">
               Action
             </AlertDialogAction>
           </AlertDialogFooter>
@@ -42,5 +42,3 @@ export function WarningAlert() {
     </AlertDialog>
   );
 }
-
-export default WarningAlert;
