+++ src/components/Interrupt/Alert/WarningAlert.tsx
@@ -11,29 +11,29 @@ import {
 } from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 
-export function SuccessAlert() {
+export function WarningAlert() {
   return (
     <AlertDialog>
       <AlertDialogTrigger asChild>
         <Button
           variant="outline"
-          className="w-fit border-green-500 text-green-500 hover:bg-green-100 hover:text-green-500"
+          className="w-fit border-amber-500 text-amber-500 hover:bg-amber-100 hover:text-amber-500"
         >
           Warning
         </Button>
       </AlertDialogTrigger>
       <AlertDialogContent className="ltr">
-        <div className="border-s-4 border-green-500 pl-6">
+        <div className="border-s-4 border-amber-500 pl-6">
           <AlertDialogHeader>
-            <AlertDialogTitle>Success</AlertDialogTitle>
+            <AlertDialogTitle>Warning Message</AlertDialogTitle>
             <AlertDialogDescription>
-              Operation completed successfully! Everything went according to
-              plan.
+              This is a warning message. Please take caution before proceeding
+              further.
             </AlertDialogDescription>
           </AlertDialogHeader>
           <AlertDialogFooter>
             <AlertDialogCancel>Dismiss</AlertDialogCancel>
-            <AlertDialogAction className="w-fit bg-green-500 text-white hover:bg-green-700 hover:text-white focus:ring-2 focus:ring-green-500">
+            <AlertDialogAction className="w-fit bg-amber-500 text-white hover:bg-amber-700 hover:text-white focus:ring-2 focus:ring-amber-500">
               Action
             </AlertDialogAction>
           </AlertDialogFooter>
@@ -42,5 +42,3 @@ export function SuccessAlert() {
     </AlertDialog>
   );
 }
-
-export default SuccessAlert;
